JBR-7459: Wayland: touchpad scrolling is too sensitive.
Part 1:
- Remaking the mapping of wl_pointer::axis events values to MouseWheelEvent rotations to eliminate the touchpad scrolling behavior "the more slowly the fingers move, the more pixels are scrolled";
- Accumulating the fraction parts of wl_pointer::axis events values to improve the accuracy of touchpad scrolling.

+++ src/java.desktop/unix/classes/sun/awt/wl/WLPointerEvent.java
@@ -1,6 +1,6 @@
 /*
- * Copyright (c) 2022, Oracle and/or its affiliates. All rights reserved.
- * Copyright (c) 2022, JetBrains s.r.o.. All rights reserved.
+ * Copyright (c) 2022-2024, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2022-2024, JetBrains s.r.o.. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -59,7 +59,7 @@ class WLPointerEvent {
     private boolean isButtonPressed; // true if button was pressed, false if released
 
     private boolean axis_0_valid; // is vertical scroll included in this event?
-    private int     axis_0_value; // "length of vector in surface-local coordinate space" (source: wayland.xml)
+    private double  axis_0_value; // "length of vector in surface-local coordinate space" (source: wayland.xml)
 
     private WLPointerEvent() {}
 
@@ -246,7 +246,7 @@ class WLPointerEvent {
         return axis_0_valid;
     }
 
-    public int getAxis0Value() {
+    public double getAxis0Value() {
         assert hasAxisEvent();
         return axis_0_value;
     }
